Refactor parsers and include them in polyglot namespace
