--- Assets/Polyglot/Scripts/CsvTsvReader/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 46773031bce16eb4fb27344d6bb14afc
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
